scripts/ml/requirements.txt
pandas==2.3.3
numpy==2.1.3
scikit-learn==1.5.2
requests==2.32.3
joblib==1.4.2
python-dotenv==1.0.1
